change int variables to double

+++ src/fr/parisnanterre/miage/poa/shapes/Point2D.java
@@ -1,17 +1,20 @@
 package fr.parisnanterre.miage.poa.shapes;
 
+/**
+ * c avec du vert => classe abstraite
+ */
 public class Point2D extends Parent implements Interface
 {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
     /**
     * Default Point2D constructor
     */
-    public Point2D(int x, int y) {
-    super();
-    this.x = x;
-    this.y = y;
+    public Point2D(double x, double y) {
+        super();
+        this.x = x;
+        this.y = y;
     }
-
+    
 }
